Products

+++ src/main/java/com/example/final_project/products/repository/ProductsRepository.java
@@ -3,6 +3,6 @@ package com.example.final_project.products.repository;
 import com.example.final_project.products.entity.Products;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductsRepository extends JpaRepository<Products, Integer> {
+public interface ProductsRepository extends JpaRepository<Products, Long> {
 
 }
